+++ ChangeLog.md
@@ -1,5 +1,8 @@
 # Change Log - oav
 
+## 09/03/2021 2.8.1
+
+- support calculate test scenario operation coverage
 
 ## 08/17/2021 2.8.0
 
@@ -11,7 +14,6 @@
 - Do not write default variable with placeholder
 - Fix validate-traffic file path resolved error.
 - Fix generate-examples can't generate some http code in response
-
 
 ## 08/12/2021 2.7.3